requirements.txt
# Core dependencies for the Multimodal DeepResearcher System

openai>=1.35.0 
aiohttp>=3.9.5 
asyncio-throttle>=1.0.2 
pydantic>=2.8.0 
pyyaml>=6.0.1 
pandas>=2.2.2 
matplotlib>=3.9.0 
seaborn>=0.13.2 
plotly>=5.22.0 
beautifulsoup4>=4.12.3 
requests>=2.32.3 
python-dotenv>=1.0.1 
markdown>=3.6.0 
jinja2>=3.1.4 
fastapi>=0.111.0 
uvicorn>=0.30.0 
redis>=5.0.0 
fastapi-limiter>=0.1.6 
pyrate-limiter>=3.6.0 
python-jose[cryptography]>=3.3.0 
passlib[bcrypt]>=1.7.4 
aiofiles>=23.2.1